0000: storage: refactor files
