+++ pkg/storage/helpers.go
@@ -1,0 +1,10 @@
+package storage
+
+func IsAlphaNumeric(s string) bool {
+	for _, r := range s {
+		if !((r >= 'A' && r <= 'Z') || (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9')) {
+			return false
+		}
+	}
+	return true
+}
